*: decode iat and exp as int64
Since #56 this package encodes the 'iat' and 'exp' claims as int64
instead of float64. However it still decodes them as a float64.
This causes comparisions to fail in tests that use some version of
reflect.DeepEqual. See coreos/dex#418.

To avoid this, attempt to decode the 'iat' and 'exp' explicitly as
int64 if possible.

+++ jose/jwt_test.go
@@ -22,7 +22,7 @@ func TestParseJWT(t *testing.T) {
 			Claims{
 				"iss": "joe",
 				// NOTE: test numbers must be floats for equality checks to work since values are converted form interface{} to float64 by default.
-				"exp": 1300819380.0,
+				"exp": int64(1300819380),
 				"http://example.com/is_root": true,
 			},
 		},
